[feat] #19 엔티티 수정 및 피드백
