+++ src/main/java/kyonggi/bookslyserver/domain/event/entity/timeEvent/TimeEventSchedule.java
@@ -1,6 +1,7 @@
-package kyonggi.bookslyserver.domain.event.entity;
+package kyonggi.bookslyserver.domain.event.entity.timeEvent;
 
 import jakarta.persistence.*;
+import kyonggi.bookslyserver.domain.event.entity.timeEvent.TimeEvent;
 import kyonggi.bookslyserver.global.common.BaseTimeEntity;
 import lombok.*;
