+++ src/main/java/kyonggi/bookslyserver/domain/event/repository/ClosingEventRepository.java
@@ -1,6 +1,6 @@
 package kyonggi.bookslyserver.domain.event.repository;
 
-import kyonggi.bookslyserver.domain.event.entity.ClosingEvent;
+import kyonggi.bookslyserver.domain.event.entity.closeEvent.ClosingEvent;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
