+++ src/main/java/kyonggi/bookslyserver/domain/event/repository/TimeEventRepository.java
@@ -1,6 +1,6 @@
 package kyonggi.bookslyserver.domain.event.repository;
 
-import kyonggi.bookslyserver.domain.event.entity.TimeEvent;
+import kyonggi.bookslyserver.domain.event.entity.timeEvent.TimeEvent;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
