+++ src/main/java/kyonggi/bookslyserver/domain/reservation/entity/ReservationSchedule.java
@@ -1,7 +1,7 @@
 package kyonggi.bookslyserver.domain.reservation.entity;
 
 import jakarta.persistence.*;
-import kyonggi.bookslyserver.domain.event.entity.ClosingEvent;
+import kyonggi.bookslyserver.domain.event.entity.closeEvent.ClosingEvent;
 import kyonggi.bookslyserver.domain.shop.entity.Shop;
 import kyonggi.bookslyserver.domain.shop.entity.WorkDays;
 import kyonggi.bookslyserver.global.common.BaseTimeEntity;
