+++ src/main/java/kyonggi/bookslyserver/domain/shop/entity/Address.java
@@ -3,9 +3,13 @@ package kyonggi.bookslyserver.domain.shop.entity;
 import jakarta.persistence.*;
 import lombok.*;
 
+import java.util.ArrayList;
+import java.util.List;
+
 @Entity
 @Getter
 @Setter
+@Builder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 @ToString(of = {"first_address", "second_address", "third_address"})
@@ -22,5 +26,7 @@ public class Address {
 
     private String third_address;
 
+    @OneToMany(mappedBy = "address", cascade = CascadeType.PERSIST)
+    private List<Shop> shops = new ArrayList<>();
 
 }
